+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- New field mappings for new table design
+- Frontend changes in new table
+
 ## [0.0.8] - 2021-11-22
 
 ### Fixed